Remove add...() and update...() functions for Attendee and Event
Use the Entry classes wherever possible! This is so that you DRY (don't repeat yourself). The Entry classes have the ReadById() implemented in $entry->__constructor(); insert() and update() implemented in $entry->save(). Using the classes abstracts away the database layer so that you need not thing about it when you want to access the database. delete() functionality is coming soon!